requirements.txt
kafka-python
pandas
scikit-learn
joblib
